controles de tipos de usuarios (admin)

+++ src/main/java/com/vpi/springboot/ControladorRest/AdministradorController.java
@@ -3,6 +3,7 @@ package com.vpi.springboot.ControladorRest;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.ConstraintViolationException;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +21,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.vpi.springboot.Logica.AdministradorService;
 import com.vpi.springboot.Modelo.Administrador;
+import com.vpi.springboot.exception.UsuarioException;
+import com.vpi.springboot.security.util.JwtUtil;
+import com.vpi.springboot.security.util.JwtUtil.keyInfoJWT;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
@@ -29,6 +33,14 @@ public class AdministradorController {
 	@Autowired
 	private AdministradorService service;
 
+	
+    @Autowired
+    private JwtUtil jwtUtil;
+    
+	@Autowired
+	private HttpServletRequest request;
+	
+
 //	@GetMapping("/getallClientes")
 //	public List<Cliente> getAllUser() {
 //		return userService.getAllClientes();
@@ -37,6 +49,9 @@ public class AdministradorController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@PostMapping("/crear")
 	public ResponseEntity<?> altaAdministrador(@RequestBody Administrador admin) {
+		if(!esAdmin()) {
+			return new ResponseEntity<>(new UsuarioException(UsuarioException.NoPermisosException("ADMIN")).getMessage(), HttpStatus.FORBIDDEN);
+		}
 		try {
 			service.crearAdministrador(admin);
 			return new ResponseEntity<Administrador>(admin, HttpStatus.OK);
@@ -50,6 +65,9 @@ public class AdministradorController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@PostMapping("/eliminar")
 	public ResponseEntity<?> eliminarUsuario(@RequestParam String mail) {
+		if(!esAdmin()) {
+			return new ResponseEntity<>(new UsuarioException(UsuarioException.NoPermisosException("ADMIN")).getMessage(), HttpStatus.FORBIDDEN);
+		}
 		try {
 			service.eliminarUsuario(mail);
 			return new ResponseEntity<String>("Eliminado correctamente", HttpStatus.OK);
@@ -61,6 +79,9 @@ public class AdministradorController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@PostMapping("/bloquear")
 	public ResponseEntity<?> bloquearUsuario(@RequestParam String mail, @RequestParam String clienteRestaurante) {
+		if(!esAdmin()) {
+			return new ResponseEntity<>(new UsuarioException(UsuarioException.NoPermisosException("ADMIN")).getMessage(), HttpStatus.FORBIDDEN);
+		}
 		try {
 			service.bloquearUsuario(mail, clienteRestaurante);
 			return new ResponseEntity<String>("Bloqueado correctamente", HttpStatus.OK);
@@ -72,6 +93,9 @@ public class AdministradorController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@PostMapping("/desbloquear")
 	public ResponseEntity<?> desbloquearUsuario(@RequestParam String mail, @RequestParam String clienteRestaurante) {
+		if(!esAdmin()) {
+			return new ResponseEntity<>(new UsuarioException(UsuarioException.NoPermisosException("ADMIN")).getMessage(), HttpStatus.FORBIDDEN);
+		}
 		try {
 			service.desbloquearUsuario(mail, clienteRestaurante);
 			return new ResponseEntity<String>("Desbloqueado correctamente", HttpStatus.OK);
@@ -99,11 +123,57 @@ public class AdministradorController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@PostMapping("/cambiarEstado/{varRestaurante}")
 	public ResponseEntity<?> cambiarEstadoRestaurante(@PathVariable String varRestaurante, @RequestParam (required = true) int estado ) {
+		if(!esAdmin()) {
+			return new ResponseEntity<>(new UsuarioException(UsuarioException.NoPermisosException("ADMIN")).getMessage(), HttpStatus.FORBIDDEN);
+		}
 		try {
 			service.cambiarEstadoRestaurante(varRestaurante, estado);
 			return new ResponseEntity<String>("Cambio de estado exitoso", HttpStatus.OK);
 		} catch (Exception e) {
 			return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
 		}
+	}
+	
+	
+	
+	
+	
+	/**
+	 * 
+	 * CONTROLES DE SEGURIDAD
+	 */
+	
+	
+	
+	
+	private Boolean esAdmin() {
+		return getInfoFromJwt(keyInfoJWT.user_type.name()).contains("ADMIN");
+	}
+
+	/**
+	 * 
+	 * @param info: mail, user_type
+	 * @return un String extraido del jwt conteniendo la info solicitada  
+	 */
+	private String getInfoFromJwt(String infoName) {
+		//obtenemos el token del header y le sacamos "Bearer "
+        final String authorizationHeader = request.getHeader("Authorization");
+
+        
+        String infoSolicitada = null;
+        String jwt = null;
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            jwt = authorizationHeader.substring(7);
+            switch(infoName){
+            case "mail":
+            	infoSolicitada = jwtUtil.extractUsername(jwt);
+            case "user_type":
+            	infoSolicitada = jwtUtil.extractUserType(jwt);
+            	
+            }
+            
+        }
+        return infoSolicitada;
 	}
 }
